+++ cb-common/src/lib.rs
@@ -62,14 +62,14 @@ pub trait App {
     id: BarId,
     device: &wgpu::Device,
     format: wgpu::TextureFormat,
-    scale: f32,
+    scale: f64,
     width: u32,
     height: u32,
   );
   fn dirty(&self, id: BarId) -> bool;
   fn move_mouse(&mut self, id: BarId, pos: Option<(f64, f64)>);
   fn click_mouse(&mut self, id: BarId, pos: (f64, f64));
-  fn set_scale(&mut self, id: BarId, device: &wgpu::Device, factor: i32);
+  fn set_scale(&mut self, id: BarId, device: &wgpu::Device, factor: f64);
   fn draw(&mut self, id: BarId, device: &wgpu::Device, queue: &wgpu::Queue, output: &wgpu::Texture);
 }
 
@@ -80,7 +80,7 @@ pub struct Bar {
   surface:        wgpu::Surface<'static>,
   surface_config: wgpu::SurfaceConfiguration,
 
-  pub scale: f32,
+  pub scale: f64,
 }
 
 impl BarId {
@@ -114,7 +114,7 @@ impl<A: App> Gpu<A> {
     &mut self,
     id: BarId,
     surface: wgpu::Surface<'static>,
-    scale: f32,
+    scale: f64,
     width: u32,
     height: u32,
   ) {
@@ -181,7 +181,7 @@ impl<A: App> Gpu<A> {
     output.present();
   }
 
-  pub fn set_scale(&mut self, id: BarId, factor: i32) {
+  pub fn set_scale(&mut self, id: BarId, factor: f64) {
     let Some(bar) = self.bars.get_mut(&id) else { return };
 
     bar.surface_config.width = bar.surface_config.width * factor as u32;
